Remove pages.ts, refactor DropGenerator
